REFACTOR: Refatorando AuthController e otimizando DespesaService
Renomeado AuthController, movido para o namespace Auth e atualizado as importações de acordo. No DespesaService, substituiu o método de recuperação do usuário por um acesso de relacionamento mais direto. Mensagem de erro aprimorada no AuthController para maior clareza.
